Move empty_candidate_inputs to common
